+++ netbox_secrets/forms/__init__.py
@@ -1,6 +1,6 @@
 from .secrets import *
 
-__all__ = (
+__all__ = [
     'validate_rsa_key',
     'SecretRoleForm',
     'SecretRoleBulkEditForm',
@@ -8,5 +8,5 @@ __all__ = (
     'SecretForm',
     'SecretFilterForm',
     'UserKeyForm',
-    'ActivateUserKeyForm'
-)
+    'ActivateUserKeyForm',
+]