+++ netbox_secrets/models/__init__.py
@@ -1,8 +1,3 @@
 from .secrets import *
 
-__all__ = (
-    'Secret',
-    'SecretRole',
-    'UserKey',
-    'SessionKey'
-)
+__all__ = ['Secret', 'SecretRole', 'UserKey', 'SessionKey']